+++ lib/domain/entities/firebase_pagination_cursor.dart
