+++ lib/domain/usecases/firestore_use_case.dart
@@ -2,7 +2,7 @@ import 'dart:io';
 
 import 'package:bread_place/domain/entities/bakery.dart';
 import 'package:bread_place/domain/entities/bakery_review_entity.dart';
-import 'package:bread_place/domain/firebase_pagination_cursor.dart';
+import 'package:bread_place/domain/entities/firebase_pagination_cursor.dart';
 import 'package:bread_place/domain/repositories/firestore_repository.dart';
 
 class FirestoreUseCase {
